Use a splay order statistic tree for the ordered factions list
This is a special type of binary search tree that reduces the lookup time for more recently accessed elements. It is also able to index each element of the tree. All operations with this should be roughly O(log n), which should be more beneficial over the previous O(n) implementation. Realtime statistics with this collection should now be feasable on insanely large servers of 10k+ active factions.